requirements.txt
fastapi
uvicorn[standard]
pydantic
scapy
websockets
pyahocorasick
SQLAlchemy
psycopg2-binary
python-dotenv
prometheus-client
pytest
pypcapkit
httptools